+++ src/datum.rs
@@ -1,6 +1,7 @@
 use crate::blockchain::Fee;
 use crate::crypto::{self, Keys, Signatures};
 use crate::role::{Accordant, Arbitrating, SwapRole};
+use crate::transaction::TxId;
 
 pub trait Datum {}
 
@@ -10,16 +11,6 @@ where
 {
     pub tx_id: TxId,
     pub tx_value: Ar::Transaction,
-}
-
-// TODO(h4sh3d): move this into transaction:: module
-pub enum TxId {
-    Funding,
-    Lock,
-    Buy,
-    Cancel,
-    Refund,
-    Publish,
 }
 
 pub struct Key<Ar, Ac>
