+++ src/lib.rs
@@ -10,5 +10,6 @@ pub mod negotiation;
 pub mod protocol;
 pub mod role;
 pub mod session;
+pub mod swap;
 pub mod transaction;
 pub mod version;
